chore(ion-identity-client): Refactor LoginService
