+++ packages/ion_identity_client/lib/src/auth/data_sources/data_sources.dart
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: ice License 1.0
 
 export 'create_recovery_credentials_data_source.dart';
-export 'login_data_source.dart';
 export 'register_data_source.dart';
