+++ packages/ion_identity_client/lib/src/auth/result_types/result_types.dart
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: ice License 1.0
 
 export 'create_recovery_credentials_result.dart';
-export 'login_user_result.dart';
 export 'recover_user_result.dart';
 export 'register_user_result.dart';
