+++ packages/ion_identity_client/lib/src/auth/services/services.dart
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: ice License 1.0
 
 export 'create_recovery_credentials_service.dart';
-export 'login_service.dart';
 export 'register_service.dart';
